Tweaks

+++ src/main/java/mrrock/com/wishlistminiproject/Controller/WishlistController.java
@@ -53,12 +53,12 @@ private UserService userService;
     //endpoint til at authenticate at oplysningerne matcher i databasen og på den måde blive logget ind
     @PostMapping("/login")
     public String requestLogIn(@RequestParam String username, @RequestParam String password, HttpSession session, RedirectAttributes redirectAttributes) {
-        User authedUser = userService.authenticateLogin(username,password);
-if (authedUser!= null){
+            User authedUser = userService.authenticateLogin(username,password);
+    if (authedUser!= null){
     session.setAttribute("user",authedUser);
     return "redirect:/overview";
-}
-redirectAttributes.addAttribute("error","Invalid Username or Password, try again");
+    }
+    redirectAttributes.addAttribute("error","Invalid Username or Password, try again");
         return "redirect:/login";
     }
 
@@ -73,6 +73,7 @@ redirectAttributes.addAttribute("error","Invalid Username or Password, try again
         redirectAttributes.addAttribute("error","Invalid Username or Password, try again");
         return "redirect:/register";
     }
+
     @GetMapping("/overview")
     public String overviewPage(HttpSession session,Model model){
         if (session.getAttribute("user")==null){
@@ -159,12 +160,22 @@ model.addAttribute("owner",userService.checkIfUserOwnsWish((User) session.getAtt
 
 
     @PostMapping("wishlist/delete/{wishlistID}")
-    public String deleteWishlist(int wishlistID){
-        userService.deleteWishlist(wishlistID);
+    public String deleteWishlist(@PathVariable int wishlistID, HttpSession session){
+        if (userService.checkIfUserOwnList(wishlistID,(User) session.getAttribute("user"))){
+            userService.deleteWishlist(wishlistID);
+        }
         return "redirect:/overview";
     }
 
 
+    @PostMapping("/delete/{wishID}")
+    public String deleteWish(@PathVariable int wishID, HttpSession session, @RequestParam int wishListID){
+        if (userService.checkIfUserOwnsWish((User) session.getAttribute("user"),wishID)){
+            userService.deleteWishFromID(wishID);
+        }
+
+        return "redirect:/wishlist/" + wishListID;
+    }
 
 
 }
